split into multi files + bios
