+++ server/routes/admin.js
@@ -1,6 +1,5 @@
 import { verifyRole } from '@/middleware';
 import { keysToCamel } from '@/common/utils';
-import { verifyRole } from '@/middleware';
 import express from 'express';
 
 import { db } from '../db/db-pgp';
